requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.2
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
python-telegram-bot==20.1
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
urllib3==1.26.14
